Updated_CMS-main/backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
supabase==2.3.4
python-dotenv==1.0.0
Werkzeug==3.0.1
requests==2.31.0
beautifulsoup4==4.12.2
google-generativeai==0.3.2
psycopg2-binary==2.9.9
python-dateutil==2.8.2
pydantic==2.5.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0.post1
xhtml2pdf>=0.2.16
reportlab>=4.0.0
fpdf==1.7.2
pandas==2.1.4
